requirements.txt
requests==2.18.4
google-auth==1.6.1